AI - Tasks/Image Captioning/requirements.txt
tensorflow==2.13.0
numpy
matplotlib
Pillow
tqdm
opencv-python
scikit-learn

